+++ src/migrations/Install.php
@@ -21,6 +21,7 @@ use putyourlightson\blitz\records\ElementQueryAttributeRecord;
 use putyourlightson\blitz\records\ElementQueryCacheRecord;
 use putyourlightson\blitz\records\ElementQueryFieldRecord;
 use putyourlightson\blitz\records\ElementQueryRecord;
+use putyourlightson\blitz\records\ElementQuerySiteRecord;
 use putyourlightson\blitz\records\ElementQuerySourceRecord;
 use putyourlightson\blitz\records\IncludeRecord;
 use putyourlightson\blitz\records\SsiIncludeCacheRecord;
@@ -51,6 +52,7 @@ class Install extends Migration
         $this->dropTableIfExists(DriverDataRecord::tableName());
         $this->dropTableIfExists(ElementQueryFieldRecord::tableName());
         $this->dropTableIfExists(ElementQueryAttributeRecord::tableName());
+        $this->dropTableIfExists(ElementQuerySiteRecord::tableName());
         $this->dropTableIfExists(ElementQuerySourceRecord::tableName());
         $this->dropTableIfExists(ElementQueryCacheRecord::tableName());
         $this->dropTableIfExists(ElementQueryRecord::tableName());
@@ -121,6 +123,14 @@ class Install extends Migration
                 'queryId' => $this->integer()->notNull(),
                 'sourceId' => $this->integer()->notNull(),
                 'PRIMARY KEY([[queryId]], [[sourceId]])',
+            ]);
+        }
+
+        if (!$this->db->tableExists(ElementQuerySiteRecord::tableName())) {
+            $this->createTable(ElementQuerySiteRecord::tableName(), [
+                'queryId' => $this->integer()->notNull(),
+                'siteId' => $this->integer()->notNull()->defaultValue(BaseDataModel::SITE_ID_ANY),
+                'PRIMARY KEY([[queryId]], [[siteId]])',
             ]);
         }
 
@@ -229,6 +239,7 @@ class Install extends Migration
         $this->addForeignKey(null, ElementQueryCacheRecord::tableName(), 'cacheId', CacheRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
         $this->addForeignKey(null, ElementQueryCacheRecord::tableName(), 'queryId', ElementQueryRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
         $this->addForeignKey(null, ElementQuerySourceRecord::tableName(), 'queryId', ElementQueryRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
+        $this->addForeignKey(null, ElementQuerySiteRecord::tableName(), 'queryId', ElementQueryRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
         $this->addForeignKey(null, ElementQueryAttributeRecord::tableName(), 'queryId', ElementQueryRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
         $this->addForeignKey(null, ElementQueryFieldRecord::tableName(), 'queryId', ElementQueryRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
         $this->addForeignKey(null, CacheTagRecord::tableName(), 'cacheId', CacheRecord::tableName(), 'id', 'CASCADE', 'CASCADE');